ERP/Maintenance_Work_Order/requirements.txt
fastapi
uvicorn
pydantic
pyjwt
bcrypt
python-dotenv
aiofiles
requests
pytest
gunicorn
python-multipart
